+++ src/graphql.rs
@@ -72,7 +72,7 @@ pub struct Query(
 );
 
 #[derive(Default, MergedObject)]
-pub struct Mutation(status::ConfigMutation);
+pub struct Mutation(status::ConfigMutation, status::SamplingPolicyMutation);
 
 #[derive(InputObject, Serialize, Clone, Debug)]
 pub struct TimeRange {